% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchoredTrueWAP.R
\name{anchoredTrueWAP}
\alias{anchoredTrueWAP}
\title{Title anchoredTrueWAP}
\usage{
anchoredTrueWAP(high, low, close, true_range, period)
}
\arguments{
\item{high}{Vector of High Values}

\item{low}{Vector of Low Values}

\item{close}{Vector of Close Values}

\item{true_range}{Vector of True Range Values}

\item{period}{Vector of bars since start of fixed period}
}
\value{
Vector of Anchored TrueWAP values
}
\description{
Calculates Anchored True Range-Weighted Average Price (TrueWAP)
}
\examples{
data(nikkei)
anchoredTrueWAP(
high = nikkei$High
, low = nikkei$Low
, close = nikkei$Close
, true_range = nikkei$tr
, period = nikkei$bars_since_segment
)
}
