% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchoredVWAP.R
\name{anchoredVWAP}
\alias{anchoredVWAP}
\title{Title anchoredVWAP}
\usage{
anchoredVWAP(HLC3, volume, period)
}
\arguments{
\item{HLC3}{Vector of High, Low, Close Average Values}

\item{volume}{Vector of Volume values}

\item{period}{Vector of bars since start of fixed period}
}
\value{
Vector of Anchored VWAP values
}
\description{
Calculates Anchored Volume-Weighted Average Price (VWAP)
}
\examples{
data(nikkei)
anchoredVWAP(
HLC3 = nikkei$HLC3
, volume = nikkei$Volume
, period = nikkei$bars_since_segment
)
}
