% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWAP.R
\name{TWAP}
\alias{TWAP}
\title{Title TWAP}
\usage{
TWAP(OHLC, period)
}
\arguments{
\item{OHLC}{Data frame object with Open, High, Low, & Close fields}

\item{period}{Rolling window length}
}
\value{
Vector of TWAP values
}
\description{
Calculates Time-Weighted Average Price (TWAP)
}
\examples{
data(nikkei)
TWAP(nikkei$OHLC, 50)
}
