% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueWAP.R
\name{TrueWAP}
\alias{TrueWAP}
\title{Title TrueWAP}
\usage{
TrueWAP(high, low, close, true_range, period)
}
\arguments{
\item{high}{Vector of High Values}

\item{low}{Vector of Low Values}

\item{close}{Vector of Close Values}

\item{true_range}{Vector of True Range Values}

\item{period}{Rolling window length}
}
\value{
Vector of TrueWAP values
}
\description{
Calculates True Range-Weighted Average Price (TrueWAP)
}
\examples{
data(nikkei)
TrueWAP(
high = nikkei$High
, low = nikkei$Low
, close = nikkei$Close
, true_range = nikkei$tr
, period = 50)
}
